test git 4

+++ src/main/java/com/example/demo/entity/Sale.java
@@ -42,7 +42,7 @@ public class Sale {
     private String describe;
 
     @Column(name = "MaximumDiscountAmount")
-    private Double maximumDiscountAmount;
+    private Integer maximumDiscountAmount;
 
     @Temporal(TemporalType.TIMESTAMP)
     @Column(name = "StartDate", nullable = false)
